Move game to package.
