+++ src/test/java/website/frontrow/ui/JBubbleBobbleUITest.java
@@ -4,7 +4,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
-import website.frontrow.Game;
+import website.frontrow.game.Game;
 import website.frontrow.board.Enemy;
 import website.frontrow.board.Player;
 import website.frontrow.board.Unit;
